Monkey patched Array class for sorting by name
- this solution is less risky
