+++ config/application.rb
@@ -12,6 +12,9 @@ module PekNext
     # Application configuration should go into files in config/initializers
     # -- all .rb files in that directory are automatically loaded.
 
+    # Custom directories with classes and modules you want to be autoloadable.
+    config.autoload_paths << Rails.root.join('lib')
+
     # Set Time.zone default to the specified zone and make Active Record auto-convert to this zone.
     # Run "rake -D time" for a list of tasks for finding time zone names. Default is UTC.
     # config.time_zone = 'Central Time (US & Canada)'
